# [Draft, don't merge] Introduce FasterHash pass - draft version