--- a/src/common/transformations/include/transformations/hash.hpp
+++ b/src/common/transformations/include/transformations/hash.hpp
@@ -34,5 +34,26 @@ private:
     uint64_t& m_hash;
 };
 
+/**
+ * @brief Hash transformation for faster calculation of hash value for ov::Model
+ */
+class NGRAPH_API FasterHash : public ov::pass::ModelPass {
+public:
+    OPENVINO_RTTI("FasterHashPass");
+
+    bool run_on_model(const std::shared_ptr<ov::Model>& f) override;
+
+    /**
+     * @brief Hash pass constructor
+     *
+     * @param output_hash_value Reference to output value. By applying hash pass on function, resulting hash value
+     * will be set to this variable
+     */
+    FasterHash(uint64_t& output_hash_value);
+
+private:
+    uint64_t& m_hash;
+};
+
 }  // namespace pass
 }  // namespace ov