### Details:
- Introduce separate 'FasterHashPass' which doesn't share any code with current serialization pass. Changes were very significant, so it was easier to introduce this. Cons: if any kind of 'new data' appear in ov::Model - FasterHashPass shall be not forgotten to be updated together with Serialize

- Main differences impacting performance are:
    - Model nodes: don't use serialization with PugiXml
    - Weights: simplify hash_combine function
    - Runtime info: don't use serialization for trivial types

Performance results (local machine):

Model with 10k nodes (Relu)
10k Relu: Regular: 27.3 FPS
10k Relu: Faster: 277.9 FPS


Model with 110k runtime info entries
100k_rt_infos: Regular: 18.1 FPS
100k_rt_infos: Faster: 35.8 FPS

Model with 400MB of weights size
400MB: Regular: 9.4 FPS
400MB: Faster: 18.1 FPS

### TODO:
 - More tests (current code coverage is only ~60%)
 - Handling of model inside model (e.g. 'If' operation or any other 'subgraph' op)

### Tickets:
 - 68936