@GlebKazantaev 2022-03-23
General comment about runtime info attributes and ov::Model hashing.
1. In OV we can distinguish two types of runtime attributes: attributes that impacts execution like Layout or PrimitivesPriority and the second type of attribute is informational, for example FusedNames or remaining attributes after transformations. So my concern is that for hashing we use both types, where the second type of attributes has huge volatility, and I assume if we add one more attribute or slightly change logic for another attribute, so resulting hashes will be different.
2. My second concern is about different facilities that we are trying not to forget to use for hashing like Input/Output/Tensor fields and so on. So I was thinking about centralized place in this facilities for hashing.